Fixed so reminders are only sent by amatsukaze
Before it didn't nick check so reminders might have been sent
from Remi_Scarlet or the like.

+++ amatsubot/amatsubot.py
@@ -1070,29 +1070,31 @@ def sanitizeNick(nick):
 
 
 def checkForReminders(userdata):
-	destination = xchat.find_context(channel='#asdf') 
+	destination = xchat.get_context() 
 	c.execute("SELECT * FROM Reminders ORDER BY ReminderDate DESC")
 	results = c.fetchall()
 	currTime = time.time()
-	for data in results:
-		remindTime = float(data[2])
-		if currTime>remindTime:
-			fromNick = data[0]
-			toNick = data[1] if sanitizeNick(fromNick) != sanitizeNick(data[1]) else "You"
-			dateAdded = float(data[3])
-			msg = data[4]
-			ID = data[5]
-			you = "yourself" if fromNick == "You" else "you"
+	activeNicks = ["amatsukaze","amatsukaze|bot"]
+	if destination.get_info("nick").lower() in activeNicks:
+		for data in results:
+			remindTime = float(data[2])
+			if currTime>remindTime:
+				fromNick = data[0]
+				toNick = data[1] if sanitizeNick(fromNick) != sanitizeNick(data[1]) else "You"
+				dateAdded = float(data[3])
+				msg = data[4]
+				ID = data[5]
+				you = "yourself" if fromNick == "You" else "you"
 
-			c.execute("DELETE FROM Reminders WHERE ID=?",(ID,))
-			conn.commit()
-			destination.command("msg "+toNick+ " \00304" + fromNick + " \00307set a reminder for " + you + " on \00304"+time.strftime("%a, %d %b %Y %H:%M:%S +0000",time.gmtime(dateAdded)))
-			destination.command("msg "+toNick+ " \00307" + msg)
-			#remind
-			pass
-		else:
-			break
-	return 1
+				c.execute("DELETE FROM Reminders WHERE ID=?",(ID,))
+				conn.commit()
+				destination.command("msg "+toNick+ " \00304" + fromNick + " \00307set a reminder for " + you + " on \00304"+time.strftime("%a, %d %b %Y %H:%M:%S +0000",time.gmtime(dateAdded)))
+				destination.command("msg "+toNick+ " \00307" + msg)
+				#remind
+				pass
+			else:
+				break
+		return 1
 
 hook = None
 hook = xchat.hook_timer(5000,checkForReminders) #call every 10 seconds
